Consolidate Settings UI into SettingsPage and update tests (#575)
